moving from postgres/go-pg to mysql/gorm
* updating tests to account for gorm mysql updates
* more updating of compiliation errors for tests
* updating int to uint for all ID's
* refactoring of tests to adopt with gorm and mysql

+++ pkg/utl/zlog/zlog.go
@@ -30,8 +30,8 @@ func (z *Log) Log(ctx echo.Context, source, msg string, err error, params map[st
 
 	params["source"] = source
 
-	if id, ok := ctx.Get("id").(int); ok {
-		params["id"] = id
+	if id, ok := ctx.Get("id").(uint); ok {
+		params["id"] = uint(id)
 		params["user"] = ctx.Get("username").(string)
 	}
